requirements.txt
attrs==20.3.0
Automat==20.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
Flask==1.1.2
Flask-WTF==0.14.3
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
gspread==3.7.0
gspread-dataframe==3.2.1
httplib2==0.19.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
lxml==4.6.2
MarkupSafe==1.1.1
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
parsel==1.6.0
Protego==0.1.16
protobuf==3.15.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
queuelib==1.5.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
Twisted==21.2.0
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.4
w3lib==1.22.0
Werkzeug==1.0.1
WTForms==2.3.3
zope.interface==5.2.0
